+++ CHANGELOG.md
@@ -22,6 +22,9 @@ Breaking Changes
 
 ### Removed
 
+- Additional signatures, other than sender and sponsor, are not supported by the tx processor in Sui. Removed
+  the additional signatures property of `SignerBlock`
+
 ## [0.17.0] - 2023-04-22
 
 Breaking Changes